Numbers_Prime Factors Sum.c
A program to find the sum of prime factors of a given integer.

It is guaranteed that every positive integer can be written as a product of its prime factors. A prime factor is itself a prime number; multiplying all the prime factors of a number should yield the number itself.